feat: add core files to projects

+++ AdoNetTask.SqlServer/dbo/Tables/Product.sql
@@ -3,8 +3,8 @@
 	[Id] INT NOT NULL PRIMARY KEY, 
     [Name] NVARCHAR(50) NOT NULL, 
     [Description] NVARCHAR(500) NULL, 
-    [Weight] INT NULL, 
-    [Height] INT NULL, 
-    [Width] INT NULL, 
-    [Length] INT NULL
+    [Weight] DECIMAL(18, 3) NULL, 
+    [Height] DECIMAL(18, 2) NULL, 
+    [Width] DECIMAL(18, 2) NULL, 
+    [Length] DECIMAL(18, 3) NULL
 )
